Refactor code, remove unwanted code
