+++ example/run.go
@@ -6,14 +6,8 @@ import (
 
 func main() {
 	server := gossamer.NewServer()
-	
+
 	server.UseStore(gossamer.NewMongoStore("localhost"))
 
 	server.Start()
-}
-
-type DefaultSensingProfileHandler struct {
-}
-
-type DefaultTaskingProfileHandler struct {
 }
