+++ request.go
@@ -1,7 +1,6 @@
 package gossamer
 
 import (
-	"log"
 	"net/url"
 	"strings"
 )
@@ -53,115 +52,21 @@ func CreateRequest(url *url.URL) (Request, error) {
 	rp.items = items
 
 	queryOpts, _ := CreateQueryOptions(url.RawQuery)
-	req := &DefaultRequest{
+	req := &GossamerRequest{
 		resourcePath: rp,
 		queryOptions: queryOpts,
 	}
 	return req, nil
 }
 
-type DefaultRequest struct {
+type GossamerRequest struct {
 	resourcePath ResourcePath
 	queryOptions QueryOptions
 }
 
-func (r *DefaultRequest) GetProtocol() ProtocolType     { return 0 }
-func (r *DefaultRequest) GetQueryOptions() QueryOptions { return nil }
+func (r *GossamerRequest) GetProtocol() ProtocolType     { return 0 }
+func (r *GossamerRequest) GetQueryOptions() QueryOptions { return nil }
 
-func (r *DefaultRequest) GetResourcePath() ResourcePath {
+func (r *GossamerRequest) GetResourcePath() ResourcePath {
 	return r.resourcePath
-}
-
-type DefaultQueryOption struct {
-	expandOption  QueryOption
-	selectOption  QueryOption
-	orderByOption QueryOption
-	topOption     QueryOption
-	skipOption    QueryOption
-	countOption   QueryOption
-	filterOption  QueryOption
-}
-
-func (o *DefaultQueryOption) Set(optType QueryOptionType, value QueryOption) {
-	switch optType {
-	case QUERYOPT_EXPAND:
-		o.expandOption = value
-
-	case QUERYOPT_COUNT:
-		o.countOption = value
-
-	case QUERYOPT_FILTER:
-		o.filterOption = value
-
-	case QUERYOPT_TOP:
-		o.topOption = value
-
-	case QUERYOPT_SKIP:
-		o.skipOption = value
-
-	case QUERYOPT_ORDERBY:
-		o.orderByOption = value
-
-	case QUERYOPT_SELECT:
-		o.selectOption = value
-
-	case QUERYOPT_UNKNOWN:
-		log.Println("Attempting to set unknown Query Option")
-		return
-	}
-}
-
-func (o *DefaultQueryOption) ExpandSet() bool {
-	return o.expandOption != nil
-}
-
-func (o *DefaultQueryOption) SelectSet() bool {
-	return o.selectOption != nil
-}
-
-func (o *DefaultQueryOption) OrderBySet() bool {
-	return o.orderByOption != nil
-}
-
-func (o *DefaultQueryOption) TopSet() bool {
-	return o.topOption != nil
-}
-
-func (o *DefaultQueryOption) SkipSet() bool {
-	return o.skipOption != nil
-}
-
-func (o *DefaultQueryOption) CountSet() bool {
-	return o.countOption != nil
-}
-
-func (o *DefaultQueryOption) FilterSet() bool {
-	return o.filterOption != nil
-}
-
-func (o *DefaultQueryOption) GetExpandOption() ExpandOption {
-	return o.expandOption.(ExpandOption)
-}
-func (o *DefaultQueryOption) GetSelectOption() SelectOption {
-	return o.selectOption.(SelectOption)
-}
-
-func (o *DefaultQueryOption) GetOrderByOption() OrderByOption {
-	return o.orderByOption.(OrderByOption)
-}
-
-func (o *DefaultQueryOption) GetTopOption() TopOption {
-	return o.topOption.(TopOption)
-}
-
-func (o *DefaultQueryOption) GetSkipOption() SkipOption {
-	return o.skipOption.(SkipOption)
-}
-
-func (o *DefaultQueryOption) GetCountOption() CountOption {
-	return o.countOption.(CountOption)
-}
-
-func (o *DefaultQueryOption) GetFilterOption() FilterOption {
-	return o.filterOption.(FilterOption)
 }
