+++ resourcepath.go
@@ -1,97 +1,5 @@
 package gossamer
 
-import (
-	"encoding/json"
-	"github.com/zenazn/goji/web"
-	"log"
-	"net/http"
-	"reflect"
-)
-
-type ResourceUrlType struct {
-	Name string `json:"name"`
-	Url  string `json:"url"`
-}
-
-func (s *DefaultServer) handleRootResource(c web.C, w http.ResponseWriter, r *http.Request) {
-	data := []ResourceUrlType{
-		{"Things", ResolveSelfLinkUrl("", ENTITY_THINGS)},
-		{"Locations", ResolveSelfLinkUrl("", ENTITY_LOCATIONS)},
-		{"Datastreams", ResolveSelfLinkUrl("", ENTITY_DATASTREAMS)},
-		{"Sensors", ResolveSelfLinkUrl("", ENTITY_SENSORS)},
-		{"Observations", ResolveSelfLinkUrl("", ENTITY_OBSERVATIONS)},
-		{"ObservedProperties", ResolveSelfLinkUrl("", ENTITY_OBSERVEDPROPERTIES)},
-		{"FeaturesOfInterest", ResolveSelfLinkUrl("", ENTITY_FEATURESOFINTEREST)},
-	}
-
-	v := &ValueList{
-		Value: data,
-	}
-	out, err := json.MarshalIndent(v, "", "  ")
-	if err != nil {
-		log.Println(err)
-	}
-	w.Write(out)
-}
-
-func (s *DefaultServer) handleGet(c web.C, w http.ResponseWriter, r *http.Request) {
-	req, err := CreateRequest(r.URL)
-	if err != nil {
-		log.Println(err)
-	}
-
-	rp := req.GetResourcePath()
-
-	result, err := s.dataStore.Query(rp)
-	if err != nil {
-
-	}
-
-	var jsonOut interface{}
-	if reflect.TypeOf(result).Kind() == reflect.Slice {
-		// Check for $count and include
-
-		jsonOut = &ValueList{
-			Value: result,
-		}
-	} else {
-		jsonOut = result
-	}
-
-	b, err := json.MarshalIndent(jsonOut, "", "  ")
-	if err != nil {
-		log.Println("Error converting to JSON")
-	}
-
-	_, err = w.Write(b)
-	if err != nil {
-		log.Println(err)
-	}
-	// http.Error(w, err.Error(), http.StatusBadRequest)
-}
-
-func (s *DefaultServer) handlePost(c web.C, w http.ResponseWriter, r *http.Request) {
-
-}
-
-func (s *DefaultServer) handlePut(c web.C, w http.ResponseWriter, r *http.Request) {
-
-}
-
-func (s *DefaultServer) handleDelete(c web.C, w http.ResponseWriter, r *http.Request) {
-
-}
-
-func (s *DefaultServer) handlePatch(c web.C, w http.ResponseWriter, r *http.Request) {
-
-}
-
-type ResourcePathItem interface {
-	GetEntity() EntityType
-	GetId() string
-	GetQueryOptions() QueryOptions
-}
-
 type SensorThingsResourcePathItem struct {
 	entityType   EntityType
 	entityId     string
