+++ response_objects.go
@@ -1,10 +1,10 @@
 package gossamer
 
-//type ResourceUrlType struct {
-//	name	string 	`json: "name"`
-//	url 	string	`json: "url"`
-//}
-
 type RootResourceResponse struct {
 	value []ResourceUrlType `json: "value"`
+}
+
+type ResourceUrlType struct {
+	Name string `json:"name"`
+	Url  string `json:"url"`
 }
